autorization check for delete and edit post

+++ backend/routes/posts.js
@@ -81,7 +81,7 @@ router.get('',(req, res, next) => {
   });
 
 
-// to edit an existing post
+// to edit an existing post - only user who create can edit it
   router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, res, next) => {
     console.log(req.file);
     let imagePath = req.body.imagePath;
@@ -97,9 +97,13 @@ router.get('',(req, res, next) => {
       imagePath: imagePath
     });
     console.log(post);
-    MongoPost.updateOne({_id : req.params.id}, post).then( result => {
-      console.log(result);
-      res.status(200).json({ message: 'Updated Successful !!'});
+    MongoPost.updateOne({_id : req.params.id, creator: req.userData.userId}, post).then( result => {
+      if(result.nModified>0){
+        res.status(200).json({ message: 'Updated Successful !!'});
+      }else{
+        res.status(401).json({ message: 'Not Autorized to Update!!'});
+      }
+
     });
 
   });
@@ -122,10 +126,17 @@ router.get('',(req, res, next) => {
 router.delete("/:id", checkAuth,(req, res, next) => {
 
     console.log(req.params.id);
-    MongoPost.deleteOne({_id : req.params.id}).then( result => {
-      console.log(result);
+    MongoPost.deleteOne({_id : req.params.id, creator: req.userData.userId}).then( result => {
+      //console.log(result);
+      if(result.n>0){
+        res.status(200).json({ message: " Post Deleted !!"});
+      }else{
+        res.status(401).json({ message: 'Not Autorized to Delete!!'});
+      }
     });
-    res.status(200).json({message : " Post Deleted !!"});
+
+
+
   });
 
   module.exports = router;
